Starts #24
Created a script that will intake an AppId and then go to the store
page.  Then scrapes the HTML, getting the price related "Classes".  Then
splices down the array to a string that contains a dollar amount with
the price of the game. @benfrisbie will need to work on making it take
AppIds from a DB and then return the price to a DB.